Always refresh the mirrored PR under MIR=true.
Currently, setting MIR to false disables refreshing the mirrored PR, but
leaving that variable to its default value of true does _not_ conversely
force the mirror to be refreshed -- it simply allows the mirror to be
created once if it does not already exist.

Here we implement a more sensible behaviour, where MIR=true always
forcefully refresh the mirror. Setting MIR to false disables the
mirroring, as before.

We also add an explicit "clean" target, which is used both internally to
ensure an existing mirror/pr.owl.gz file is ignored (under MIR=true),
and to give users the possibility to manually clean the mirror
directory.